start.py
import uvicorn
from dotenv import load_dotenv
import os

load_dotenv()

if __name__ == "__main__":
    uvicorn.run(
        "src.app:app",
        host="0.0.0.0",
        port=int(os.getenv("SERVER_PORT", "8000")),
        reload=True,
    )
