pyproject.toml
[project]
name = "ai-goofish-monitor"
version = "0.1.0"
description = "闲鱼智能抓取机器人"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.2",
    "fastapi[standard]>=0.135.2",
    "openai>=2.30.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = ["coverage>=7.13.5", "pytest>=9.0.2"]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "live: real traffic smoke tests that require real credentials and external services",
    "live_slow: slower optional live smoke tests such as AI task generation",
]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
